+++ src/main/java/uz/pdp/controller/AuthController.java
@@ -1,0 +1,27 @@
+package uz.pdp.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import uz.pdp.dto.SignInRequest;
+import uz.pdp.dto.SignUpRequest;
+import uz.pdp.service.AuthService;
+
+@RestController
+@RequestMapping("/api/auth")
+public class AuthController {
+
+    @Autowired
+    private AuthService authService;
+
+    @PostMapping("/sign-in")
+    public ResponseEntity<?> signIn(@Valid @RequestBody SignInRequest signInRequest) {
+        return authService.signIn(signInRequest);
+    }
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<?> signUp(@Valid @RequestBody SignUpRequest signUpRequest) {
+        return authService.signUp(signUpRequest);
+    }
+}