+++ src/main/java/uz/pdp/dto/SignInRequest.java
@@ -1,0 +1,17 @@
+package uz.pdp.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SignInRequest {
+    @NotBlank(message = "Name is required")
+    private String name;
+
+    @NotBlank(message = "Password is required")
+    private String password;
+}