+++ src/main/java/uz/pdp/entity/Company.java
@@ -5,8 +5,9 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uz.pdp.role.Socials;
 import uz.pdp.entity.Address;
+import uz.pdp.enums.Socials;
+
 import java.time.LocalDate;
 import java.util.List;
 @Data