+++ src/main/java/uz/pdp/repository/UserRepository.java
@@ -6,5 +6,6 @@ import uz.pdp.entity.User;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
+    Optional<User> findByName(String name);
+    Optional<User> findByEmail(String email);
 }